requirements.txt
certifi==2022.9.24
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
filelock==3.8.0
fonttools==4.38.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.12.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.10.1
hyperframe==5.2.0
idna==2.10
joblib==1.2.0
kiwisolver==1.4.4
libretranslatepy==2.1.1
lxml==4.9.1
matplotlib==3.6.2
nltk==3.7
numpy==1.23.4
packaging==21.3
Pillow==9.3.0
psycopg2==2.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
rfc3986==1.5.0
scipy==1.9.3
six==1.16.0
sniffio==1.3.0
tokenizers==0.13.2
torch==1.13.0
tqdm==4.64.1
transformers==4.24.0
translate==3.6.1
typing-extensions==4.4.0
urllib3==1.26.12
wordcloud==1.8.2.2
